heading->turn in brain.decide and bond.advance
the brain can't direct a target heading to the body (which works on
actual heading, which the brain doesn't know about). instead, the brain
comes up with a turn based on its believed current heading and desired
new heading, and then the body will execute that turn against the actual
heading.